pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "dj-database-url>=2.3.0",
    "django>=5.1.7",
    "django-bootstrap5>=25.1",
    "django-filter>=25.1",
    "gunicorn>=23.0.0",
    "i18n>=0.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "rollbar>=1.2.0",
    "ruff>=0.11.2",
]
